% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wTGS}
\alias{wTGS}
\title{auxiliary function that does weighted Tempered Gibbs Sampling from the posterior distribution of Gamma}
\usage{
wTGS(X_, y_, n_, p_, n_iter, burnin_, h1_, h2_, c_, k_, weighted)
}
\arguments{
\item{X_}{a matrix of p regressors (independent variables) with dimension (nxp).}

\item{y_}{a vector of n observations (dependent variable) with dimensions (nx1).}

\item{n_}{number of observations n.}

\item{p_}{number of regressors p.}

\item{n_iter}{a positive integer specifying the number of iterations for the Markov Chain.}

\item{burnin_}{a positive integer specifying the number of burn in iterations for the Markov Chain.}

\item{h1_}{first parameter of the beta distribution defining h. If value of h fixed, make it h1.}

\item{h2_}{second parameter of the beta distribution defining h. If value of h fixed, make h2=0.}

\item{c_}{constant of proportionality to prior covariace matrix.}

\item{k_}{k_weighted parameter for weighted sampling.}

\item{weighted}{boolean indicating if weighted sampling.}
}
\value{
List with values for PIP and required elements to reproduce the samples.
}
\description{
Performs weighted Tempered Gibbs Sampling
}
